requirements.txt
python-dotenv
flask
mysql-connector-python>=8.0.33